requirements.txt
asgiref==3.7.2
Django==4.2.2
et-xmlfile==1.1.0
joblib==1.2.0
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
patsy==0.5.3
Pillow==9.5.0
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post5
sqlparse==0.4.4
statsmodels==0.14.0
threadpoolctl==3.1.0
typing_extensions==4.6.3
tzdata==2023.3
